0.0.1 +++ Initial Release
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/radialsearch/trunk@80484 735d13b6-9817-0410-8766-e36946ffe9aa

+++ lib/class.tx_radialsearch_em.php
@@ -257,6 +257,7 @@ class tx_radialsearch_em
     $data       = $_POST['data'];
     $file       = $data[ 'database.selectbox' ];
     $pid        = $data[ 'database.pid' ];
+    $truncate   = ( boolean ) $data[ 'database.truncate' ];
     
       // RETURN : no file selected
     if( empty( $file ) )
@@ -304,8 +305,11 @@ class tx_radialsearch_em
     
     $rows = array( );
     
-    $query = 'TRUNCATE tx_radialsearch_postalcodes';
-    $res    = $GLOBALS['TYPO3_DB']->sql_query( $query ); 
+    if( $truncate )
+    {
+      $query = 'TRUNCATE tx_radialsearch_postalcodes';
+      $res    = $GLOBALS['TYPO3_DB']->sql_query( $query ); 
+    }
 
     $keys = 'INSERT INTO typo3_browser.tx_radialsearch_postalcodes ( uid, pid, tstamp, crdate, cruser_id, deleted, country_code, postal_code, place_name, admin_name1, admin_code1, admin_name2, admin_code2, admin_name3, admin_code3, latitude, longitude, accuracy ) VALUES ' . PHP_EOL;
     $i = 0;
